added search jsx
