+++ src/components/Navigation/Navigation.module.scss
@@ -1,77 +1,1 @@
-.itemDetails{
-  width: 350px;
-  position: fixed;
-  z-Index: 1001;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  padding: 40px;
-  background: rgba(14, 22, 36, 0.91);
-  border: 3px solid rosybrown;
-}
 
-
-.list {
-  display: flex;
-  gap: 40px;
-  padding: 0;
-}
-
-.item {
-  list-style: none;
-}
-
-.searchResults{
-  background: rgba(25, 24, 24, 0.84);
-  border-radius: 5px;
-  position: absolute;
-  width: 350px;
-  padding: 10px;
-  z-index: 1000;
-  max-height: 54vh;
-  overflow-y: auto;
-}
-
-.title{
-  color: #00ff40;
-  font-size: 18px;
-  text-align: center;
-  margin-top: 10px;
-  text-transform: uppercase;
-}
-
-.desc {
-  color: white;
-  margin-top: 6px;
-  font-size: 14px;
-  text-align: center;
-}
-
-.weight, .price {
-  color: #fea6a6;
-  margin-top: 6px;
-  text-align: center;
-}
-
-.categories {
-  color: #e40096;
-  text-align: center;
-  border: 1px solid;
-  max-width: 175px;
-  margin: 6px auto;
-  border-radius: 5px;
-}
-
-.searchResult {
-  color: #5d9ae9;
-  &:hover {
-    color: #fea6a6;
-    cursor: pointer;
-  }
-}
-
-@media (max-width: 355px) {
-  .list {
-    gap: 15px;
-  }
-}
